requirements.txt
leancloud-sdk>=1.0.9,<=2.0.0
Flask==0.12
Flask-Sockets>=0.1,<1.0
Jinja2==2.9.6
Werkzeug==0.12.1
gunicorn==19.6.0
gevent>=1.0.2,<2.0.0
gevent-websocket>=0.9.5,<1.0.0